requirements.txt
customtkinter==5.2.1
pymongo==4.6.1
requests==2.31.0
pywin32==306; sys_platform == 'win32'
Pillow==10.1.0
aiohttp==3.9.1
PyInstaller==6.3.0